+++ CHANGELOG.md
@@ -630,6 +630,7 @@ Initial release
 ### [defmt-decoder-next]
 
 * [#1073] Gather symbols from unmerged `.defmt.*` sections, don't use `DW_AT_location`
+* [#1066] Add support for biased, wrapped index decoding
 
 ### [defmt-decoder-v1.1.0] (2026-01-20)
 
@@ -1043,6 +1044,7 @@ Initial release
 [#1083]: https://github.com/knurling-rs/defmt/pull/1083
 [#1073]: https://github.com/knurling-rs/defmt/pull/1073
 [#1070]: https://github.com/knurling-rs/defmt/pull/1070
+[#1066]: https://github.com/knurling-rs/defmt/pull/1066
 [#1068]: https://github.com/knurling-rs/defmt/pull/1068
 [#1053]: https://github.com/knurling-rs/defmt/pull/1053
 [#1055]: https://github.com/knurling-rs/defmt/pull/1055